fea: to_max_feature
